Move some Method from [Domain.UnitTests] to [Framework.DataType.UnitTests]
